Start DP190Medium
I got pretty far, I think. The only problem is that it is taking a ton
of time to finish. I don't know why. Maybe I should use a faster search.
Now that I think of it, that's probably the issue. When I get back,
I'll write a binary search or something else nice.
Add a comment.