It mostly works
There's one that disagrees with the problem, and maybe it should be
that way, but I think I could make the case that my way is reasonable.
It certainly makes sense, and it works for all the other ones. Now I'm
going to try to make it pretty by moving code all over the place. We'll
see how that goes.